Add support for custom update (#7)
* Enable custom_updater.
